Migrated templatedComponent to mgt-element
