+++ packages/mgt/src/components/mgt-get/mgt-get.ts
@@ -6,10 +6,8 @@
  */
 
 import { customElement, html, property } from 'lit-element';
-import { Providers, ProviderState } from '@microsoft/mgt-element';
+import { Providers, ProviderState, MgtTemplatedComponent, equals } from '@microsoft/mgt-element';
 import { prepScopes } from '../../utils/GraphHelpers';
-import { equals } from '../../utils/Utils';
-import { MgtTemplatedComponent } from '../templatedComponent';
 
 /**
  * Custom element for making Microsoft Graph get queries
