+++ packages/mgt/src/components/mgt-login/mgt-login.ts
@@ -8,11 +8,9 @@
 import { customElement, html, property } from 'lit-element';
 import { classMap } from 'lit-html/directives/class-map';
 import { IDynamicPerson } from '../../graph/types';
-import { Providers, ProviderState } from '@microsoft/mgt-element';
+import { Providers, ProviderState, MgtTemplatedComponent } from '@microsoft/mgt-element';
 import { MgtFlyout } from '../sub-components/mgt-flyout/mgt-flyout';
-import { MgtTemplatedComponent } from '../templatedComponent';
 import { styles } from './mgt-login-css';
-
 import { getUserWithPhoto } from '../../graph/graph.user';
 import '../../styles/fabric-icon-font';
 import '../mgt-person/mgt-person';
